Bug Fix, duration should be float, so there was a bad comparison using a silent cast to float.

+++ include/waveform.h
@@ -11,19 +11,19 @@ namespace RustyAudio
 class Waveform
 {
 public:
-    Waveform(unsigned int duration) :
+    Waveform(float duration) :
         mDuration(duration)
     {}
     
     virtual std::int32_t operator()(float milliseconds) const { return 0.0f; };
 
-    unsigned int duration() const
+    float duration() const
     {
         return mDuration;
     }
 
 protected:
-    unsigned int mDuration;
+    float mDuration;
 };
 
 }
